backend/requirements.txt
asgiref==3.9.1
charset-normalizer==3.4.3
Django==5.2.5
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==11.3.0
psycopg2-binary==2.9.9
PyJWT==2.10.1
reportlab==4.4.3
sqlparse==0.5.3
django-cors-headers==4.3.0
django-environ==0.12.0
django-allauth==65.11.1
requests==2.32.5
cryptography==45.0.7
django-cors-headers==4.8.0